// src/gen-ui/lib/core/base-types.ts
export enum Axis {
  horizontal = "horizontal",
  vertical = "vertical",
}

export enum Clip {
  none = "none",
  /**
   * # 使用硬边裁剪
   */
  hardEdge = "hardEdge",
  /**
   * # 使用抗锯齿裁剪
   */
  antiAlias = "antiAlias",
}
export enum MainAxisAlignment {
  start = "start",
  end = "end",
  center = "center",
  spaceBetween = "spaceBetween",
  spaceAround = "spaceAround",
  spaceEvenly = "spaceEvenly",
}

export enum CrossAxisAlignment {
  start = "start",
  end = "end",
  center = "center",
  stretch = "stretch",
  baseline = "baseline",
}
// ...
export enum MainAxisSize {
  /**
   * # 最大尺寸
   */
  max = "max",
  /**
   * # 最小尺寸
   */
  min = "min",
}